ds-sc-32/requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
dnspython==2.1.0
idna==3.2
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
py-openaq==1.1.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
